WIP update

+++ Examples/3_RealWorld/1_TerrainGen/1_PlaneGen.cpp
@@ -55,7 +55,7 @@ int main() {
         renderer.flush();
 
         Coel::update(window);
-        player.update(window.elapsed);
+        player.update(static_cast<float>(window.elapsed));
     }
 
     Coel::destroy(shader);
